requirements.txt
streamlit
pypdf
scikit-learn
requests
python-dotenv
